v.0.5.0.9 fixes in frmImages. Added Happy Birthday. Split container in frmTopics.
frmImages: put under try catch a strange behaviour when saving with a new text of . Why it behaves in that fashoin was NOT understood.
Added Happy Birthday functionality: when it is the Birthsday of a students, the window with the student's data is open and "Happy Birthday" is played. .
Added horizontal split Container in frmTopics.
Removed the unhooking of events from TreeMptt, because the behaviour was akward. .