Deep revision of TreeMptt looking for a subtle error THAT I DID NOT FIND!
I am almost sure that the possible concurrency errors between the background and foreground savings have been corrected.
Anyway we have a bad and disruptive error that stops the programs when we are modifying the TreeView component in in every form that uses it.
Version number increased to 1.57.3
TreeMptt:
MpttLeftRight: added new class (to refactor in future, maybe as a partial class) to hold the methods to save the tree in background without using an UI component.
TreeMpttDb: added method SaveOnlyLeftAndRightNodes that saves left and right pointer, leaving all the info in Topic like it was.
CommonsWinForms: added precautions in switching the LED of background saving.
frmSetup: added safety to concurrent acces during the writing of config file
Topic class: added the clone method, used during the catch of the error in synchronization between the background task and the foreground UI.
  Now it isn't used  but I leave it here because it could be useful in the future .